requirements.txt
# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Database
datasette>=0.64.0
sqlite-utils>=3.35.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Configuration
PyYAML>=6.0.0

# Utilities
tqdm>=4.66.0
python-slugify>=8.0.0

# Development
pytest>=7.4.0
black>=23.0.0
